Sort `mdbook` redirects
There were a few comments, but I think it’s better to make it all
uniform and avoid these comments. The comment was about the v2
rewrite, but this is now done and so the comment is less useful.

I also normalized the quotes to double-quotes instead of
single-quotes. The latter is a literal string in TOML, but we don’t
actually have any special characters to escape here.